feat: cancel the context if the program panics and add a global `Flush` function to IO
